fix(pageSize): refactored to use the right components and local storage
