+++ components/molecules/table/components/wrapper.js
@@ -4,14 +4,16 @@
 
 // React
 import React, { memo, useEffect, useRef } from 'react'
-import { array, bool, func, number, oneOfType, shape, string } from 'prop-types'
+
+// Style
+import styled from 'styled-components'
 
 // UI
 import { TableContent, TableLoading, TablePagination } from '../../../'
 import { tableColumnCount } from '../utils/columnCount'
 
-// Style
-import styled, { css } from 'styled-components'
+// Props
+import { defaultProps, propTypes } from '../props'
 
 export const Table = memo(
   ({
@@ -38,13 +40,17 @@ export const Table = memo(
     const tableRef = useRef(null)
     const tableReady = useRef(false)
     const rowLength = rows.length > 0
+
     useEffect(() => {
       const table = tableRef.current
+
       if (table && loading !== true && tableReady.current === false) {
         tableReady.current = true
       }
+
       return () => {}
     }, [rowLength])
+
     return (
       <StyledWrapper fullHeight={fullHeight} isLoading={loading} border={border}>
         <TableLoading colsLength={tableSpan} show={loading} />
@@ -90,14 +96,15 @@ const StyledWrapper = styled.div`
   ${({ fullHeight, isLoading }) => (fullHeight || isLoading !== undefined) && 'height: 100%;'}
   position: relative;
   width: 100%;
+
   ${({ isLoading }) =>
     isLoading &&
-    css`
+    `
       position: relative;
     `}
   ${({ border }) =>
     border === false &&
-    css`
+    `
       td,
       th {
         border: none !important;
@@ -108,7 +115,7 @@ const StyledWrapper = styled.div`
 const StyledResponsive = styled.div`
   ${({ responsive }) =>
     responsive &&
-    css`
+    `
       background-color: #fff;
       display: block;
       overflow-x: auto;
@@ -116,44 +123,5 @@ const StyledResponsive = styled.div`
     `}
 `
 
-Table.propTypes = {
-  align: oneOfType([string, bool]),
-  border: bool,
-  caption: string,
-  className: string,
-  columns: array,
-  fullHeight: bool,
-  hover: bool,
-  loading: bool,
-  pagination: bool,
-  paginationProps: shape({
-    currentPage: number,
-    onPageChange: func,
-    onPageSizeChange: func,
-    perPage: number
-  }),
-  responsive: bool,
-  rowClick: func,
-  rows: array.isRequired,
-  setSort: func,
-  sort: shape({
-    item: string,
-    order: string
-  }),
-  striped: bool
-}
-
-Table.defaultProps = {
-  align: false,
-  border: true,
-  columns: [],
-  className: 'Table',
-  fullHeight: false,
-  hover: true,
-  loading: undefined,
-  noData: true,
-  pagination: false,
-  paginationProps: {},
-  responsive: true,
-  striped: true
-}
+Table.propTypes = propTypes
+Table.defaultProps = defaultProps
